requirements.txt
flask
requests
kafka-python
redis
reedsolo
